Report test re-runs with 'rerun' field
Gerrit verify-status plugin used to track test reruns by looking for
a special "recheck" message in the category field.  Now reruns are
tracked with a special rerun bit.  Decoupling the two provides for
a more flexible solution.  We can now report to Gerrit using the rerun
field instead of using the category field.