Package gerrit-rest-java-client into this plugin
We override RevisionApi classes in gerrit-rest-java-client to include
the verifystatus endpoint.  When run using `java -jar jenkins.war` the
jenkins plugin classpath puts the gerrit-rest-java-client ahead of the
RevisionApi class that's in this plugin which results in a
NoSuchMethodError[1].  To make jenkins use this version of RevisionApi
we repackage gerrit-rest-java-client into this plugin excluding the
RevisionApi classes.

[1] java.lang.NoSuchMethodError: com.google.gerrit.extensions.api.changes
.RevisionApi.verifyStatus()Lcom/google/gerrit/extensions/api/changes/VerifyStatusApi;